pyproject.toml
[build-system]
requires = ["setuptools>=64.0.1", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "compositionspace"
dynamic = ["version"]
authors = [
    { name = "Alaukik Saxena, Sarath Menon, Mariano Forti, Markus Kühbach" },
]
description = "APT analysis tools"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "h5py",
    "scikit-learn",
    "trimesh",
    "ifes_apt_tc_data_modeling>=0.2.1",  
    "pyyaml",
    "flatdict",
    "lxml",
    "jupyter",
    "jupyterlab",
    "jupyterlab_h5web",
]

[project.urls]
"Homepage" = "https://github.com/eisenforschung/CompositionSpace"

[project.optional-dependencies]
dev = [
    "mypy",
    "ruff",
    "pandas",
    "matplotlib",
    "tqdm",
    "pyevtk",
    "pyvista",
    "pytest",
    "pytest-timeout",
    "pytest-cov",
    "pip-tools",
]

[tool.setuptools_scm]
version_scheme = "no-guess-dev"
local_scheme = "node-and-date"

[tool.setuptools.packages.find]
include = ["compositionspace/*"]
exclude = ["development/*", "nexus_definitions/*"]

[tool.ruff]
include = ["compositionspace/*.py", "tests/*.py"]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "W", # pycodestyle
    "PL", # pylint
]
ignore = [
    "E501", # Line too long ({width} > {limit} characters)
    "E701", # Multiple statements on one line (colon)
    "E731", # Do not assign a lambda expression, use a def
    "E402",  # Module level import not at top of file
    "PLR0911", # Too many return statements
    "PLR0912", # Too many branches
    "PLR0913", # Too many arguments in function definition
    "PLR0915", # Too many statements
    "PLR2004", # Magic value used instead of constant
    "PLW0603", # Using the global statement
    "PLW2901", # redefined-loop-name
    "PLR1714", # consider-using-in
    "PLR5501", # else-if-used
]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
strict = false
ignore_missing_imports = true
follow_imports = "silent"
no_strict_optional = true
disable_error_code = "import, annotation-unchecked"
